board,comment,post,notice API
커밋합니다.

+++ backend/src/main/java/com/univercity/unlimited/greenUniverCity/function/notice/repository/NoticeRepository.java
@@ -3,6 +3,5 @@ package com.univercity.unlimited.greenUniverCity.function.notice.repository;
 import com.univercity.unlimited.greenUniverCity.function.notice.entity.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeRepository extends JpaRepository <Notice,Integer>{
-    //uservo의 타입
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 }
